--- a/test/unit/components/html-message.tsx
+++ b/test/unit/components/html-message.tsx
@@ -122,7 +122,7 @@ describe('<FormattedHTMLMessage>', () => {
     ).find('span');
 
     expect((rendered.prop('dangerouslySetInnerHTML') as any).__html).toBe(
-      'Hello, <b>[object Object]</b>!'
+      'Hello, <b></b>!'
     );
   });
 });