Newer versions has new AST & better perf

Before
```
<div> x 2,028,960 ops/sec ±3.31% (83 runs sampled)
100 x <div/> x 10,623 ops/sec ±4.06% (77 runs sampled)
100 x <FormattedNumber> x 1,418 ops/sec ±7.83% (76 runs sampled)
100 x <FormattedDate> x 692 ops/sec ±3.88% (76 runs sampled)
100 x <FormattedMessage> x 2,866 ops/sec ±3.96% (89 runs sampled)
100 x <FormattedMessage> with placeholder x 56.80 ops/sec ±12.36% (51 runs sampled)
100 x <FormattedMessage> with placeholder in AST form x 110 ops/sec ±2.44% (70 runs sampled)
100 x <FormattedMessage> with placeholder, cached x 473 ops/sec ±9.41% (63 runs sampled)
100 x <FormattedMessage> with placeholder, cached in AST form x 119 ops/sec ±6.87% (56 runs sampled)
```

After
```
<div> x 2,111,875 ops/sec ±2.44% (90 runs sampled)
100 x <div/> x 12,008 ops/sec ±1.31% (89 runs sampled)
100 x <FormattedNumber> x 1,982 ops/sec ±0.99% (91 runs sampled)
100 x <FormattedDate> x 830 ops/sec ±2.12% (86 runs sampled)
100 x <FormattedMessage> x 2,981 ops/sec ±0.92% (88 runs sampled)
100 x <FormattedMessage> with placeholder x 76.83 ops/sec ±9.83% (69 runs sampled)
100 x <FormattedMessage> with placeholder in AST form x 206 ops/sec ±1.12% (79 runs sampled)
100 x <FormattedMessage> with placeholder, cached x 573 ops/sec ±1.07% (84 runs sampled)
100 x <FormattedMessage> with placeholder, cached in AST form x 392 ops/sec ±4.28% (82 runs sampled)
```